Adds convertion from old gene panels to new ones
Old gene panels where based on hgnc symbols which is a unstable gene identifier.
We have decided to use hgnc id:s as identifiers so we need to convert all the old panels to this new format.
The command `scout convert` does exactly that. Genes that have ambigous identifiers will be collapsed, these
needs manual work to choose the correct gene.

Cleans some of the command files.